Add project skeleton.
The chosen base framework is Lumen, Laravel was not quite right for the
requirements since the project meant to be only an API.

Lumen is a micro-framework with all the architectural benefits of
Laravel, with all the good parts like Eloquent, caching, queues,
validation, routing, middleware, service container, etc.

This code struct is meant to be unit testable, scalable and well
documented, and organized.